+++ gems/kettle-jem/spec/thin_slice_spec.rb
@@ -11172,6 +11172,7 @@ RSpec.describe Kettle::Jem do
   it "renders README dev and test stack table with self-exclusion" do
     example_table = described_class.send(:readme_dev_test_stack_table, "example")
     turbo_table = described_class.send(:readme_dev_test_stack_table, "turbo_tests2")
+    coverage_table = described_class.send(:readme_dev_test_stack_table, "kettle-soup-cover")
 
     expect(example_table).to include("[appraisal2](https://bestgems.org/gems/appraisal2)")
     expect(example_table).to include("[GitHub](https://github.com/appraisal-rb/appraisal2)")
@@ -11179,10 +11180,14 @@ RSpec.describe Kettle::Jem do
     expect(example_table).to include("[turbo_tests2](https://bestgems.org/gems/turbo_tests2)")
     expect(example_table).to include("[kettle-test](https://bestgems.org/gems/kettle-test)")
     expect(example_table).to include("[GitHub](https://github.com/kettle-dev/kettle-test)")
+    expect(example_table).to include("[kettle-soup-cover](https://bestgems.org/gems/kettle-soup-cover)")
+    expect(example_table).to include("[GitHub](https://github.com/kettle-dev/kettle-soup-cover)")
     expect(example_table).to include("[rubocop-lts](https://bestgems.org/gems/rubocop-lts)")
     expect(turbo_table).not_to include("[turbo_tests2](https://bestgems.org/gems/turbo_tests2)")
     expect(turbo_table).to include("[appraisal2](https://bestgems.org/gems/appraisal2)")
     expect(turbo_table).to include("[kettle-test](https://bestgems.org/gems/kettle-test)")
+    expect(coverage_table).not_to include("[kettle-soup-cover](https://bestgems.org/gems/kettle-soup-cover)")
+    expect(coverage_table).to include("[kettle-test](https://bestgems.org/gems/kettle-test)")
   end
 
   it "wires Appraisal2 RuboCop as a generator plugin without generated appraisal leakage" do